clean up the code

+++ routes/users.js
@@ -1,34 +1,18 @@
 /**
  * TODO:
  * - создать контролер
- * --------------------
  * - привести в порядок роутер
- * - удалить lodash за ненадобностью
- * - добавить в редактирвание защиту, проверка айди с токена и фронта
  */
 
 let express = require('express');
 let router = express.Router();
 let bcrypt = require('bcryptjs');
 let User = require('../models/users.models');
-require('dotenv').config();
-// для защиты от инъекций используем lodash ("вырезает лишнее")
-let _ = require('lodash');
-let passport = require('passport');
-let jwt = require('jsonwebtoken');
-let multer = require('multer');
-let uniqid = require('uniqid');
+//require('dotenv').config();
 let fs = require('fs');
-
-function sendJSONresponse(res, status, content) {
-    console.log(status, content);
-    res.status(status);
-    res.json(content);
-}
-
-function compare(a, b) {
-    return (a === b);
-}
+let fmain = require('../functions/fmain');
+let fauth = require('../functions/fauth');
+let upload = require('../functions/fupload-ava');
 
 function getIp(req) {
     let ip = req.headers['x-forwarded-for'] ||
@@ -36,27 +20,6 @@ function getIp(req) {
         req.socket.remoteAddress ||
         (req.connection.socket ? req.connection.socket.remoteAddress : null);
     return ip.replace(/::ffff:/, '');
-}
-
-function checkAuth(req, res, next) {
-    passport.authenticate('jwt', { session: false }, (err, decryptToken, jwtError) => {
-        if (jwtError != void (0) || err != void (0)) {
-            //console.log('err = ', err);
-            //console.log('jwtError = ', jwtError);
-            sendJSONresponse(res, 403, { error: err || jwtError });
-            return;
-        }
-        req.user = decryptToken;
-        next();
-    })(req, res, next);
-}
-
-function createToken(body) {
-    return jwt.sign(
-        body,
-        process.env.secretOrKey,
-        { expiresIn: '1 day' }
-    );
 }
 
 function userData(user) {
@@ -71,76 +34,18 @@ function userData(user) {
     };
 }
 
-let storage = multer.diskStorage({
-    destination: function (req, file, cb) {
-        cb(null, './public/images/avatars/');
-    },
-
-    filename: function (req, file, cb) {
-        if (req.params && req.params.id) {
-            User
-                .findById(_.escapeRegExp(req.params.id))
-                .exec(function (err, user) {
-
-                    if (!user) {
-                        cb(new Error(err));
-                        return;
-                    } else if (err) {
-                        cb(new Error(err));
-                        return;
-                    }
-
-                    let ext;
-                    file.origFileName = user.avatar;
-
-                    let newFileName = uniqid() + Date.now();
-
-                    switch (file.mimetype) {
-                        case 'image/jpeg':
-                            ext = '.jpg';
-                            cb(null, newFileName + ext);
-                            break;
-                        case 'image/pjpeg':
-                            ext = '.pjpeg';
-                            cb(null, newFileName + ext);
-                            break;
-                        case 'image/png':
-                            ext = '.png';
-                            cb(null, newFileName + ext);
-                            break;
-                        case 'image/gif':
-                            ext = '.gif';
-                            cb(null, newFileName + ext);
-                            break;
-                        default:
-                            cb(new Error('You can only upload files with the extension jpg, png, gif'));
-                            return;
-                    }
-                });
-        } else {
-            cb(new Error('no id in request'));
-        }
-
-
-    }
-});
-
-let upload = multer({
-    storage: storage,
-    limits: { fileSize: 1024 * 1024 } // Filesize avatar max. 1 MB
-}).single('avatar');
 
 // login user (Done)
 router.post('/login', function (req, res) {
     if (req.body && req.body.email && req.body.password) {
         User
-            .findOne({ email: { $regex: _.escapeRegExp(req.body.email.toLowerCase()), $options: "i" } }, function (err, user) {
+            .findOne({ email: req.body.email.toLowerCase() }, function (err, user) {
                 if (err) {
-                    sendJSONresponse(res, 400, err);
+                    fmain.sendJSONresponse(res, 400, err);
                     return;
                 }
-                if (user && bcrypt.compareSync(req.body.password, user.password) && compare(user.email, req.body.email.toLowerCase())) {
-                    let token = createToken({
+                if (user && bcrypt.compareSync(req.body.password, user.password) && (user.email === req.body.email.toLowerCase())) {
+                    let token = fauth.createToken({
                         id: user._id,
                         role: user.role,
                         username: user.name,
@@ -153,16 +58,16 @@ router.post('/login', function (req, res) {
                     let userPlusToken = userData(user);
                     userPlusToken.token = token;
 
-                    sendJSONresponse(res, 200, userPlusToken);
+                    fmain.sendJSONresponse(res, 200, userPlusToken);
                 } else {
-                    sendJSONresponse(res, 400, {
+                    fmain.sendJSONresponse(res, 400, {
                         message: "Wrong login or password"
                     });
                     return;
                 }
             });
     } else {
-        sendJSONresponse(res, 400, {
+        fmain.sendJSONresponse(res, 400, {
             "message": "All required fields must be filled"
         });
     }
@@ -171,7 +76,7 @@ router.post('/login', function (req, res) {
 // logout user
 router.post('/logout', (req, res) => {
     res.clearCookie('token');
-    sendJSONresponse(res, 200, {
+    fmain.sendJSONresponse(res, 200, {
         "message": "logout success"
     });
 });
@@ -180,13 +85,13 @@ router.post('/logout', (req, res) => {
 router.post('/users', function (req, res) {
     if (req.body && req.body.email && req.body.password && req.body.phoneNumber && req.body.role) {
         User
-            .findOne({ email: { $regex: _.escapeRegExp(req.body.email), $options: "i" } }, function (err, user) {
+            .findOne({ email: req.body.email }, function (err, user) {
                 if (err) {
-                    sendJSONresponse(res, 400, err);
+                    fmain.sendJSONresponse(res, 400, err);
                     return;
                 }
                 if (user) {
-                    sendJSONresponse(res, 400, {
+                    fmain.sendJSONresponse(res, 400, {
                         message: "User with email or phone number already exist"
                     });
                     return;
@@ -205,14 +110,14 @@ router.post('/users', function (req, res) {
                     ip: getIp(req)
                 }, function (err, user) {
                     if (err) {
-                        sendJSONresponse(res, 400, err);
+                        fmain.sendJSONresponse(res, 400, err);
                     } else {
-                        sendJSONresponse(res, 201, userData(user));
+                        fmain.sendJSONresponse(res, 201, userData(user));
                     }
                 });
             });
     } else {
-        sendJSONresponse(res, 400, {
+        fmain.sendJSONresponse(res, 400, {
             "message": "All required fields must be filled"
         });
     }
@@ -223,7 +128,7 @@ router.get('/users', function (req, res) {
     User.find({})
         .exec(function (err, users) {
             if (err) {
-                sendJSONresponse(res, 404, err);
+                fmain.sendJSONresponse(res, 404, err);
                 return;
             }
             if (users) {
@@ -231,9 +136,9 @@ router.get('/users', function (req, res) {
                 users.forEach(element => {
                     usersOut.push(userData(element));
                 });
-                sendJSONresponse(res, 200, usersOut);
+                fmain.sendJSONresponse(res, 200, usersOut);
             } else {
-                sendJSONresponse(res, 404, {
+                fmain.sendJSONresponse(res, 404, {
                     message: "users not found"
                 });
             }
@@ -245,33 +150,32 @@ router.get('/users', function (req, res) {
 router.get('/users/:id', function (req, res) {
     if (req.params && req.params.id) {
         User
-            .findById(_.escapeRegExp(req.params.id))
+            .findById(req.params.id)
             .exec(function (err, user) {
                 if (!user) {
-                    sendJSONresponse(res, 404, {
+                    fmain.sendJSONresponse(res, 404, {
                         message: "User not found"
                     });
                     return;
                 } else if (err) {
-                    sendJSONresponse(res, 404, err);
+                    fmain.sendJSONresponse(res, 404, err);
                     return;
                 }
-                sendJSONresponse(res, 200, userData(user));
+                fmain.sendJSONresponse(res, 200, userData(user));
             });
     } else {
-        sendJSONresponse(res, 404, {
+        fmain.sendJSONresponse(res, 404, {
             message: "no id in request"
         });
     }
 });
 
 // Modify User (Done)
-router.put('/users/:id', checkAuth, function (req, res) {
-    console.log(req.user);
-    if (req.params && req.params.id) {
+router.put('/users/:id', fauth.checkAuth, function (req, res) {
+    if (req.params && req.params.id && (req.params.id === req.user.id)) {
         upload(req, res, function (err) {
             if (err) {
-                sendJSONresponse(res, 500, err);
+                fmain.sendJSONresponse(res, 500, err);
                 return;
             }
 
@@ -288,34 +192,35 @@ router.put('/users/:id', checkAuth, function (req, res) {
             User
                 .findByIdAndUpdate(req.params.id, user, { new: true }, function (err, user) {
                     if (!user) {
-                        sendJSONresponse(res, 404, err);
+                        fmain.sendJSONresponse(res, 404, err);
                         return;
                     } else if (err) {
-                        sendJSONresponse(res, 500, err);
+                        fmain.sendJSONresponse(res, 500, err);
                         return;
                     }
 
 
-                    if (
-                        (fs.existsSync(req.file.destination + req.file.origFileName))
-                        && (req.file.origFileName !== 'def_customer.jpg')
-                        && (req.file.origFileName !== 'def_master.jpg')
-                    ) {
-                        fs.unlinkSync(req.file.destination + req.file.origFileName, (err) => {
-                            if (err) console.log(req.file.origFileName + ' not deleted');
-                        });
-                    } else {
-                        console.log(req.file.destination + req.file.origFileName + ' not found');
+                    if (req.file) {
+                        if (
+                            (fs.existsSync(req.file.destination + req.file.origFileName))
+                            && (req.file.origFileName !== 'def_customer.jpg')
+                            && (req.file.origFileName !== 'def_master.jpg')
+                        ) {
+                            fs.unlinkSync(req.file.destination + req.file.origFileName, (err) => {
+                                if (err) console.log(req.file.origFileName + ' not deleted');
+                            });
+                        } else {
+                            console.log(req.file.destination + req.file.origFileName + ' not found');
+                        }
                     }
 
-
-                    sendJSONresponse(res, 200, userData(user));
+                    fmain.sendJSONresponse(res, 200, userData(user));
 
                 });
         });
     } else {
-        sendJSONresponse(res, 404, {
-            message: "no id in request"
+        fmain.sendJSONresponse(res, 404, {
+            message: "no id in request or bad token"
         });
     }
